+++ lib/cogment/client_actor.cpp
@@ -72,4 +72,6 @@ void Client_actor::dispatch_observation(const cogment::Observation& obs, bool en
 
 void Client_actor::dispatch_reward(int /*tick_id*/, const ::cogment::Reward& /*reward*/) {}
 
+void Client_actor::dispatch_message(int /*tick_id*/, const ::cogment::Message& /*message*/) {}
+
 }  // namespace cogment