+++ lib/cogment/client_actor.h
@@ -40,6 +40,7 @@ class Client_actor : public Actor {
 
   void dispatch_observation(const cogment::Observation& obs, bool end_of_trial) override;
   void dispatch_reward(int tick_id, const ::cogment::Reward& reward) override;
+  void dispatch_message(int tick_id, const ::cogment::Message& message) override;
 
   bool is_active() const override;