+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Orchestrator support messages, they can be sent between actors and the environment.
+
 ## v1.0.0-alpha1 - 2020-12-07
 
 - Initial alpha release, expect some breaking changes.